Move type to avoid import cycles and duplicate
* move openapi var from api to openapi.go
* move GeojsonFeatureData, FeatureCollection, MakeGeojsonFeatureJSON to api.go (remove duplicates)
* move Table, Column, Function, TransformFunction, Sorting from catalog.go to api.go
* move JSONType and PGType to datatype.go

Split and regroup type struct from api.go

* create api/function.go and api/table.go
* regroup type struct by 'domain' in api.go
