+++ internal/service/handler_test.go
@@ -132,7 +132,7 @@ func TestCollectionItem(t *testing.T) {
 func TestFilterB(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?prop_b=1")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -142,7 +142,7 @@ func TestFilterB(t *testing.T) {
 func TestFilterD(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_c/items?prop_d=1")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -152,7 +152,7 @@ func TestFilterD(t *testing.T) {
 func TestFilterBD(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_c/items?prop_b=2&prop_d=2")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -162,7 +162,7 @@ func TestFilterBD(t *testing.T) {
 func TestFilterBDNone(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_c/items?prop_b=1&prop_d=2")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -172,7 +172,7 @@ func TestFilterBDNone(t *testing.T) {
 func TestSortBy(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?sortby=prop_b")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -182,7 +182,7 @@ func TestSortBy(t *testing.T) {
 func TestSortByDesc(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?sortby=-prop_b")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -192,7 +192,7 @@ func TestSortByDesc(t *testing.T) {
 func TestSortByAsc(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?sortby=+prop_b")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -202,7 +202,7 @@ func TestSortByAsc(t *testing.T) {
 func TestLimit(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?limit=3")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -215,7 +215,7 @@ func TestLimit(t *testing.T) {
 func TestLimitZero(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?limit=0")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -230,7 +230,7 @@ func TestLimitInvalid(t *testing.T) {
 func TestQueryParamCase(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?LIMIT=2&Offset=4")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -242,7 +242,7 @@ func TestQueryParamCase(t *testing.T) {
 func TestOffset(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?limit=2&offset=4")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -285,7 +285,7 @@ func TestProperties(t *testing.T) {
 	// - non-existing names are ignored
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?limit=2&properties=PROP_A,prop_C,prop_a,not_prop")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -299,7 +299,7 @@ func TestProperties(t *testing.T) {
 func TestPropertiesAll(t *testing.T) {
 	rr := hTest.DoRequest(t, "/collections/mock_a/items?limit=2")
 
-	var v util.FeatureCollection
+	var v api.FeatureCollection
 	errUnMarsh := json.Unmarshal(hTest.ReadBody(rr), &v)
 	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
 
@@ -412,7 +412,7 @@ func checkLink(tb testing.TB, link *api.Link, rel string, conType string, href s
 	util.Equals(tb, href, link.Href, "Link href")
 }
 
-func checkFunctionSummary(tb testing.TB, v *api.FunctionSummary, fun *data.Function) {
+func checkFunctionSummary(tb testing.TB, v *api.FunctionSummary, fun *api.Function) {
 	util.Equals(tb, fun.Name, v.Name, "Function name")
 	util.Equals(tb, fun.Description, v.Description, "Function description")
 
@@ -427,7 +427,7 @@ func checkFunctionSummary(tb testing.TB, v *api.FunctionSummary, fun *data.Funct
 	}
 	checkLink(tb, v.Links[2], api.RelItems, itemsType, hTest.UrlBase+pathItems)
 }
-func checkFunction(t *testing.T, fun *data.Function) {
+func checkFunction(t *testing.T, fun *api.Function) {
 	path := "/functions/" + fun.ID
 	resp := hTest.DoRequest(t, path)
 	body, _ := ioutil.ReadAll(resp.Body)
